Create Guessing Game
To start there would be a spot to insert the guess where it should prompt to enter 1-5. The number would have already been generated as well so once the guess is submitted then it will let them know 1 of 3 things: that it is correct, it is to low, or it is to high. From there if it is correct then it will output correct and then end. If the guess is one of the other options then it will go back and allow for a retry until it is correct and then end.